requirement.txt
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.6
django-cors-headers==2.1.0
django-crispy-forms==1.6.1
django-filter==1.0.4
django-formtools==2.1
django-guardian==1.4.9
django-redis==4.8.0
django-reversion==2.0.10
djangorestframework==3.7.0
djangorestframework-jwt==1.11.0
drf-extensions==0.3.1
future==0.16.0
httplib2==0.10.3
idna==2.6
itypes==1.1.0
Jinja2==2.9.6
Markdown==2.6.9
MarkupSafe==1.0
mysqlclient==1.3.12
olefile==0.44
Pillow==4.3.0
pycryptodome==3.4.7
PyJWT==1.5.3
PyMySQL==0.7.11
pytz==2017.2
raven==6.2.1
redis==2.10.6
requests==2.18.4
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
XlsxWriter==1.0.2
xlwt==1.3.0
social-auth-app-django
